store_test.go
package main

import (
	"bytes"
	"testing"
)

func TestPathTransformFunc(t *testing.T) {
	key := "momsbestpicture"
	pathKey := CASPathTransformFunc(key)
	expectedOriginalKey := "6804429f74181a63c50c3d81d733a12f14a353ff"
	expectedPathName := "68044/29f74/181a6/3c50c/3d81d/733a1/2f14a/353ff"
	if pathKey.Pathname != expectedPathName {
		t.Errorf("Expected %s, got %s", pathKey.Pathname, expectedPathName)
	}
	if pathKey.Original != expectedOriginalKey {
		t.Errorf("Expected %s, got %s", pathKey.Original, expectedOriginalKey)
	}
}

func TestStore(t *testing.T) {
	opts := StoreOpts{
		PathTransformFunc: CASPathTransformFunc,
	}
	s := NewStore(opts)

	data := bytes.NewReader([]byte("somejpegbytes"))
	if err := s.writeStream("somespecialpicture", data); err != nil {
		t.Fatal(err)
	}
}
